Login endpoints

+++ lib/services/auth_service.dart
@@ -1,5 +1,7 @@
 import 'package:get/get.dart';
+import 'package:onesignal_flutter/onesignal_flutter.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:yht_ticket/models/responses/login_response.dart';
 import 'package:yht_ticket/routes/app_pages.dart';
 import 'package:yht_ticket/shared/constants/storage.dart';
 
@@ -10,28 +12,55 @@ class AuthService extends GetxService {
   final isLoggedIn = false.obs;
   bool get isLoggedInValue => isLoggedIn.value;
 
+  LoginResponse? loginResponse;
+
   @override
   void onInit() {
-    var token = Get.find<SharedPreferences>().getString(StorageConstants.token);
-    isLoggedIn.value = token?.isNotEmpty ?? false;
+    var storage =
+        Get.find<SharedPreferences>().getString(StorageConstants.loginResponse);
+    isLoggedIn.value = storage?.isNotEmpty ?? false;
+
+    if (isLoggedInValue) {
+      loginResponse = LoginResponse.fromJson(storage!);
+    }
+
     super.onInit();
   }
 
-  void login(String token) {
+  void login(LoginResponse loginResponse) {
     isLoggedIn.value = true;
 
-    if (token.isNotEmpty) {
-      final prefs = Get.find<SharedPreferences>();
-      prefs.setString(StorageConstants.token, token);
+    this.loginResponse = loginResponse;
 
-      Get.offAndToNamed(Routes.DASHBOARD);
-    }
+    final prefs = Get.find<SharedPreferences>();
+
+    prefs.setString(StorageConstants.loginResponse, loginResponse.toJson());
+
+    // Setting External User Id with Callback Available in SDK Version 3.9.3+
+    OneSignal.shared.setExternalUserId(loginResponse.username).then((results) {
+      print(results.toString());
+    }).catchError((error) {
+      print(error.toString());
+    });
+
+    Get.offAndToNamed(Routes.DASHBOARD);
   }
 
   void logout() {
-    isLoggedIn.value = false;
-    final prefs = Get.find<SharedPreferences>();
-    prefs.clear();
-    Get.offAndToNamed(Routes.LOGIN);
+    if (isLoggedInValue) {
+      isLoggedIn.value = false;
+
+      loginResponse = null;
+
+      OneSignal.shared.removeExternalUserId().then((results) {
+        print(results.toString());
+      }).catchError((error) {
+        print(error.toString());
+      });
+
+      final prefs = Get.find<SharedPreferences>();
+      prefs.clear();
+      Get.offAndToNamed(Routes.LOGIN);
+    }
   }
 }
